requirements.txt
google-generativeai
openai==1.20.0
pandas==2.2.2
protobuf==4.25.3
pydantic==2.7.0
Pympler==1.0.1
click==8.1.7
ollama
fastapi
uvicorn
torch
transformers
accelerate
anthropic
python-dotenv
litellm
pyfzf
npm
platformdirs
arxiv
llama-index-embeddings-huggingface
watchdog
chromadb
llama_index==0.10.19
llama_index_core==0.10.19
numpy==1.24.3